enclave/server/requirements.txt
asn1crypto==1.4.0
cryptography==36.0.1
pyOpenSSL==22.0.0
cbor2==5.4.2.post1
cose==0.9.dev8
requests==2.27.1
jsonschema==4.4.0
boto3==1.20.51
pyzipper==0.3.5
pikepdf==4.4.1
reportlab==3.6.6
twilio==7.7.1
sendgrid==6.9.7
phonenumbers==8.12.35